requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
spacy==3.5.0
Werkzeug==2.3.0
python-dotenv==1.0.0
